+++ CHANGELOG.rst
@@ -1,6 +1,20 @@
-1.0.2
+1.0.3
 =====
 In Progress
+
+Features
+--------
+* Support static columns in schemas, which are available starting in
+  Cassandra 2.1. (github issue #91)
+
+Other
+-----
+* Don't ignore column names when parsing typestrings.  This is needed for
+  user-defined type support.  (github issue #90)
+
+1.0.2
+=====
+March 4, 2014
 
 Bug Fixes
 ---------